requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
#dj-rest-auth==2.2.2
Django==3.2.11
django-allauth==0.47.0
#django-rest-auth==0.9.5
#djangorestframework==3.13.1
#djangorestframework-jwt==1.11.0
google-api-core==2.4.0
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
oauthlib==3.1.1
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
